need front-side formatting

+++ src/main/webapp/admin.jsp
@@ -6,16 +6,17 @@
 <%@ taglib prefix="sql" uri="http://java.sun.com/jstl/sql" %>
 <%@ page import="java.util.ArrayList"%>
 <html lang="en">
-<html>
     <head>
         <meta charset="utf-8">
         <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+         <script type="text/javascript" src="js/myscripts.js"></script>
         <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
     </head>
           <body>
           <div align="right">
-                  <h3><c:out value="${firstName} ${lastName}"/></h3>
-                  <h4><a href="/logout">Logout</a><h4>
+                  <h3><c:out value="${login}"/></h3>
+                  <h3><a href="/logout">Logout</a><h3>
+
               </div>
               <div>
               <a href="/create">
@@ -24,9 +25,9 @@
               </div>
               <table id="mytable" class="table table-striped" border="1">
                 <thead>
-                    <th>id</th>
+
                     <th>Login</th>
-                    <th>Password</th>
+
                     <th>First Name</th>
                     <th>Last Name</th>
                     <th>Age</th>
@@ -38,13 +39,13 @@
                 <tbody>
                  <c:forEach items="${users}" var="name">
                      <tr>
-                       <td><c:out value="${name.id}"/></td>
+
                        <td><c:out value="${name.login}"/></td>
-                       <td><c:out value="${name.password}"/></td>
+
                        <td><c:out value="${name.firstName}"/></td>
                        <td><c:out value="${name.lastName}"/></td>
                        <td><c:out value="${name.age}"/></td>
-                       <td><c:out value="${name.role_id}"/></td>
+                       <td><c:out value="${name.rolename}"/></td>
                        <td>
 
                        <c:url value="/edit" var="url" scope="request"></c:url>
@@ -57,12 +58,14 @@
                                                		<c:param name="id" value="${name.id}" />
                                                	</c:url>
 
-                                              <a href="/delete?logintodelete=<c:out value="${name.login}"/>">
+                                              <a href="/delete?logintodelete=<c:out value="${name.login}"/>" onclick="return confirmDelete();">
                                                <img border="0"  src="images/trash.jpg" width="30" height="30">
                                                </a>
                      </tr>
                    </c:forEach>
                 </tbody>
               </table>
+
+
           </body>
 </html>
